Adds working directory to create_job, moves create_job outside of Scheduler
