+++ meeshkan/core/job.py
@@ -6,6 +6,9 @@ import uuid
 import datetime
 import os
 from pathlib import Path
+import uuid
+
+from .config import JOBS_DIR
 
 
 LOGGER = logging.getLogger(__name__)
@@ -47,9 +50,17 @@ class Executable(object):
         """
         return
 
+    @staticmethod
+    def to_full_path(args: Tuple[str, ...], cwd: Path):
+        """Given args, iterates over arg and prepends .sh and .py files with given current working directory.
+        :param args
+        """
+        supported_file_suffixes = [".py", ".sh"]
+        return [cwd.joinpath(arg) if os.path.splitext(arg)[1] in supported_file_suffixes else arg for arg in args]
+
 
 class ProcessExecutable(Executable):
-    def __init__(self, args: Tuple[str, ...], output_path: Path = None):
+    def __init__(self, args: Tuple[str, ...], cwd: Path, output_path: Path = None):
         """
         Executable executed with `subprocess.Popen`.
         :param args: Command-line arguments to execute, fed into `Popen(args, ...)` _after_ prepending cwd to files
@@ -68,7 +79,8 @@ class ProcessExecutable(Executable):
                 return os.path.abspath(arg)
             return arg
 
-        self.args = [to_full_path_if_known_file(arg) for arg in args]
+        # self.args = [to_full_path_if_known_file(arg) for arg in args]
+        self.args = self.to_full_path(args, cwd)
         self.popen = None  # type: Optional[subprocess.Popen]
         self.output_path = output_path
 
@@ -185,3 +197,23 @@ class Job(object):
                 'name': self.name,
                 'status': self.status.name,
                 'args': str(self.executable)}
+
+
+def _verify_python_executable(args: Tuple[str, ...]):
+    """Simply checks if the first argument's extension is .py, and if so, prepends 'python' to args"""
+    if len(args) > 0:    # pylint: disable=len-as-condition
+        if os.path.splitext(args[0])[1] == ".py":
+            args = ("python",) + args
+    return args
+
+
+def create_job(args: Tuple[str, ...], cwd: str, job_number: int, name: str = None, poll_interval: int = Non):
+    """Creates a job from given arguments"""
+    cwd = Path(cwd)  # Convert to Path object
+    job_uuid = uuid.uuid4()
+    args = _verify_python_executable(args)
+    LOGGER.debug("Creating job for %s", args)
+    output_path = JOBS_DIR.joinpath(str(job_uuid))
+    executable = ProcessExecutable(args, output_path=output_path)
+    job_name = name or "Job #{job_number}".format(job_number=job_number)
+    return Job(executable, job_number=job_number, job_uuid=job_uuid, name=job_name, poll_interval=poll_interval)
